// Frontend Test/src/app/models/log.model.ts
export interface Log {
        cd_cebroker_state: string;
        pro_cde: number;
        cd_profession: string;
        id_license: string;
        dt_end: string;
        cd_environment: string;
        dt_Start_Log: string;
        ds_compl_status_returned: string;
        dt_end_log: string;
        cd_machine: string;
        id_client_nbr: string;
}